Fix incorrect check using reflect.Float32 instead of reflect.Float64

+++ netquery/orm/encoder.go
@@ -181,7 +181,7 @@ func DatetimeEncoder(loc *time.Location) EncodeFunc {
 				return nil, false, fmt.Errorf("failed to parse time as RFC3339: %w", err)
 			}
 
-		case (normalizedKind == reflect.Int || normalizedKind == reflect.Uint || normalizedKind == reflect.Float32) && colDef.IsTime:
+		case (normalizedKind == reflect.Int || normalizedKind == reflect.Uint || normalizedKind == reflect.Float64) && colDef.IsTime:
 			t = time.Unix(val.Int(), 0)
 
 		default:
